Code cleanup

+++ src/main/java/org/hampelratte/svdrp/sorting/RecordingLengthComparator.java
@@ -41,6 +41,6 @@ public class RecordingLengthComparator implements Comparator<Recording> {
 
     @Override
     public int compare(Recording r1, Recording r2) {
-        return (int) Math.signum(r1.getDuration() - r2.getDuration());
+        return (int) Math.signum(r1.getDuration() - (float)r2.getDuration());
     }
 }
